feat: allow status for creating build container
**Context**

Before a build executes its first command, Screwdriver needs to
perform a series of setup and configuration operations. As it is
today, there isn't a specific state for describing these pre-build
operations.

For this commit, we aim to not solve the entire problem of
describing every operation. We instead want to describe one state
of those setup operations.

**Objective**

Enable a "creating container" build status in order to describe
that a build is constructing the build container.