Add '.', ':', to IMAGE_ALIAS regex
This change will allow for an alias such as `xcode:14.1`. Without this
change the alias needs to be something unnatural such as `xcode-14_1`.

The image aliases are essentially only dictionary keys, so it's not
clear why their naming need be restricted at all, much less be any more
restrictive than the image name itself. However, on the assumption that
there is a reason for the naming restriction, this commit take the
conservative approach of adding only two additional characters.